cleanup mui ssr
