--- packages/vulcan-emotion/README.md
@@ -1,1 +1,0 @@
-Styled components package.
