--- packages/vulcan-emotion/lib/client/main.js
@@ -1,2 +1,0 @@
-import './wrapWithEmotion';
-export * from '../modules/index';
