--- packages/vulcan-emotion/lib/server/main.js
@@ -1,4 +1,0 @@
-import setupEmotion from './setupEmotion';
-setupEmotion();
-
-export * from '../modules/index';
