+++ packages/vulcan-ui-material/lib/server/main.js
@@ -1,2 +1,2 @@
 export * from '../modules/index';
-import './wrapWithMuiTheme';
+import './wrapWithMui';
